3375 - ODP References: separate component to EditorWYSIWYG/
