+++ src/client/components/EditorWYSIWYG/EditorWYSIWYGLinks/props.ts
@@ -1,0 +1,4 @@
+export type Props = {
+  onChange: (value?: string) => void
+  value: string
+}
