+++ src/client/components/EditorWYSIWYG/index.ts
@@ -1,1 +1,2 @@
 export { default } from './EditorWYSIWYG'
+export { default as EditorWYSIWYGLinks } from './EditorWYSIWYGLinks'
