+++ src/client/pages/OriginalDataPoint/components/DataSources/References/References.tsx
@@ -1,36 +1,33 @@
-import React, { useCallback, useState } from 'react'
+import React, { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { Objects } from 'utils/objects'
 
+import { OriginalDataPoint } from 'meta/assessment'
 import { Topics } from 'meta/messageCenter'
 
 import { DataCell } from 'client/components/DataGrid'
-import EditorWYSIWYG from 'client/components/EditorWYSIWYG'
+import { EditorWYSIWYGLinks } from 'client/components/EditorWYSIWYG'
 import MarkdownPreview from 'client/components/MarkdownPreview'
 import ReviewIndicator from 'client/components/ReviewIndicator'
-import { Props } from 'client/pages/OriginalDataPoint/components/DataSources/References/Props'
 import { useShowReviewIndicator } from 'client/pages/OriginalDataPoint/hooks/useShowReviewIndicator'
 
 import { useIsDisabled } from '../hooks/useIsDisabled'
 import { useUpdateDataSources } from '../hooks/useUpdateDataSources'
-import { useEditorOptions } from './hooks/useEditorOptions'
-import { useOnClose } from './hooks/useOnClose'
-import AddFromRepository from './AddFromRepository'
 
 type OnChange = (value?: string) => void
+
+type Props = {
+  originalDataPoint: OriginalDataPoint
+}
 
 const References: React.FC<Props> = (props: Props) => {
   const { originalDataPoint } = props
   const { t } = useTranslation()
 
-  const [isOpen, setIsOpen] = useState<boolean>(false)
-  const [editor, setEditor] = useState(null)
-
   const reviewIndicator = useShowReviewIndicator(originalDataPoint)
   const disabled = useIsDisabled(originalDataPoint)
-  const onClose = useOnClose({ setIsOpen, setEditor, editor })
-  const editorOptions = useEditorOptions({ setIsOpen, setEditor })
+
   const updateOriginalDataPoint = useUpdateDataSources()
 
   const onChange = useCallback<OnChange>(
@@ -52,16 +49,7 @@ const References: React.FC<Props> = (props: Props) => {
           </div>
         )}
 
-        {!disabled && (
-          <>
-            <EditorWYSIWYG
-              onChange={onChange}
-              options={editorOptions}
-              value={originalDataPoint.dataSourceReferences ?? ''}
-            />
-            <AddFromRepository isOpen={isOpen} onClose={onClose} />
-          </>
-        )}
+        {!disabled && <EditorWYSIWYGLinks onChange={onChange} value={originalDataPoint.dataSourceReferences ?? ''} />}
       </DataCell>
 
       {reviewIndicator && (
